pyproject.toml
[tool.poetry]
package-mode = false
name = "omnicore-ai"
version = "0.1.0"
description = "Voice AI SaaS platform boilerplate"
authors = ["Antigravity <antigravity@google.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.3"
vapi-server-sdk = "^0.1"
vapi-python = "^0.1"
requests = "^2.32"
google-api-python-client = "^2.188.0"
google-auth = "^2.48.0"
google-auth-oauthlib = "^1.2.4"
google-auth-httplib2 = "^0.2.0"
httpx = "^0.28.1"
aiogram = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "httpx (>=0.28.1,<0.29.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
